[Iceberg]Refactor test classes to avoid them growing too large
