+++ presto-iceberg/src/test/java/com/facebook/presto/iceberg/nessie/TestIcebergNessieCatalogDistributedQueries.java
@@ -13,38 +13,25 @@
  */
 package com.facebook.presto.iceberg.nessie;
 
-import com.facebook.presto.iceberg.IcebergDistributedTestBase;
 import com.facebook.presto.iceberg.IcebergQueryRunner;
+import com.facebook.presto.iceberg.TestIcebergDistributedQueries;
 import com.facebook.presto.testing.QueryRunner;
 import com.facebook.presto.testing.containers.NessieContainer;
 import com.google.common.collect.ImmutableMap;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
-
-import java.io.File;
-import java.util.Map;
 
 import static com.facebook.presto.iceberg.CatalogType.NESSIE;
 import static com.facebook.presto.iceberg.nessie.NessieTestUtil.nessieConnectorProperties;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-@Test
-public class TestIcebergDistributedNessie
-        extends IcebergDistributedTestBase
+public class TestIcebergNessieCatalogDistributedQueries
+        extends TestIcebergDistributedQueries
 {
     private NessieContainer nessieContainer;
 
-    protected TestIcebergDistributedNessie()
+    protected TestIcebergNessieCatalogDistributedQueries()
     {
         super(NESSIE);
-    }
-
-    @Override
-    protected Map<String, String> getProperties()
-    {
-        File metastoreDir = getCatalogDirectory();
-        return ImmutableMap.of("warehouse", metastoreDir.toString(), "uri", nessieContainer.getRestApiUri());
     }
 
     @BeforeClass
@@ -73,11 +60,8 @@ public class TestIcebergDistributedNessie
     }
 
     @Override
-    public void testExpireSnapshotWithDeletedEntries()
+    protected boolean supportsViews()
     {
-        // Nessie do not support expire snapshots as it set table property `gc.enabled` to `false` by default
-        assertThatThrownBy(() -> super.testExpireSnapshotWithDeletedEntries())
-                .isInstanceOf(RuntimeException.class)
-                .hasMessageMatching("Cannot expire snapshots: GC is disabled .*");
+        return false;
     }
 }
